+++ config/routing.php
@@ -6,7 +6,9 @@ use App\Controller\ShowUser;
 use App\Controller\ShowAllUsers;
 use App\Controller\ShowAllQuestions;
 use App\Controller\addUser;
+use App\Controller\checkUser; 
 use App\Controller\inscription;
+use App\Controller\connexion; 
 
 use Framework\Routing\Route;
 
@@ -17,6 +19,8 @@ return [
     new Route('GET', '/allUsers', ShowAllUsers::class),
     new Route('GET', '/allQuestions', ShowAllQuestions::class),
     new Route('GET', '/addUser', addUser::class),
+    new Route('GET', '/checkUser', checkUser::class),
     new Route('GET', '/inscription', inscription::class),
+    new Route('GET', '/connexion', connexion::class),
 
 ];