+++ src/App/Controller/connexion.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Controller;
+
+use Framework\Controller\AbstractController;
+
+class Connexion extends AbstractController
+{
+    public function __invoke(): string
+    {
+        return $this->render('user/connexion.html.twig');
+    }
+}